requirements.txt
torch>=2.0.0
torchvision>=0.15.0
numpy>=1.21.0
Pillow>=9.0.0
safetensors>=0.3.0
transformers>=4.21.0
opencv-python>=4.5.0
requests>=2.25.0
aiohttp>=3.8.0
websockets>=10.0
PyYAML>=6.0
tqdm>=4.64.0
functions-framework
google-cloud-storage 